efficientnet-b0_server/requirements.txt
efficientnet_pytorch==0.7.1
numpy==1.24.1
Pillow==9.5.0
Requests==2.28.2
transformers==4.28.1
